content/fingerprint-based-blood-group-detection/requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.24.3
pillow==10.4.0
tensorflow==2.10.1
